+++ tools/friendsList/tool.js
@@ -51,7 +51,8 @@
         'friendsNumberTitle:zero': 'Friends',
         'friendFilterTitle': 'Name filter',
         'opponentsText': 'Opponents',
-        'opponentsTitle': 'LiChess Tools - players you\'ve played against most'
+        'opponentsTitle': 'LiChess Tools - players you\'ve played against most',
+        'unblockPlayerTitle': 'Unblock'
       },
       'ro-RO': {
         'onlineFriends': '%s prieteni online',
@@ -81,7 +82,8 @@
         'friendsNumberTitle:zero': 'Prieteni',
         'friendFilterTitle': 'Filtru nume',
         'opponentsText': 'Adversari',
-        'opponentsTitle': 'LiChess Tools - juc\u0103tori \u00eempotriva c\u0103rora ai jucat mai mult'
+        'opponentsTitle': 'LiChess Tools - juc\u0103tori \u00eempotriva c\u0103rora ai jucat mai mult',
+        'unblockPlayerTitle': 'Deblocare'
       }
     }
 
@@ -366,6 +368,7 @@
       if (!this.options.liveFriendsPage) return;
       if (!this.isLivePage) return;
       const isFavoritesOrBlocksOrFollowers = !this.isFriendsPage;
+      const isBlocks = lt.isBlockedPlayersPage();
       if (lt.global.document.hidden) {
         lt.global.requestAnimationFrame(this.updateFriendsPage);
         return;
@@ -402,33 +405,37 @@
                     .attr('placeholder',trans.noarg('friendFilterTitle'))
                     .on('input',this.filterFriends)
                  )
-          .append($(`<i data-icon="${lt.icon.toEntity(lt.icon.Antichess)}" data-role="hideInactive">`)
+          .append($(`<i data-role="hideInactive">`)
+            .attr('data-icon',lt.icon.Antichess)
             .attr('title', trans.noarg('hideInactiveTitle'))
             .on('click', () => {
               $('main').toggleClass('lichessTools-hideInactive');
               this.scrollIfNeeded();
             }))
-          .append($(`<i data-icon="${lt.icon.toEntity(lt.icon.DiscOutline)}" data-role="hideOffline">`)
+          .append($(`<i data-role="hideOffline">`)
+            .attr('data-icon',lt.icon.DiscOutline)
             .attr('title', trans.noarg('hideOfflineTitle'))
             .on('click', () => {
               $('main').toggleClass('lichessTools-hideOffline');
               this.scrollIfNeeded();
             }))
-          .append($(`<i data-icon="${lt.icon.toEntity(lt.icon.AnalogTv)}" data-role="hideNotPlaying">`)
+          .append($(`<i data-role="hideNotPlaying">`)
+            .attr('data-icon',lt.icon.AnalogTv)
             .attr('title', trans.noarg('hideNotPlayingTitle'))
             .on('click', () => {
               $('main').toggleClass('lichessTools-hideNotPlaying');
               this.scrollIfNeeded();
-            }));
+            }))
+          .appendTo(header);
         if (!isFavoritesOrBlocksOrFollowers) {
           liveButtons
-            .append($(`<i data-icon="${lt.icon.toEntity(lt.icon.BellOutline)}" data-role="hideMuted">`)
+            .append($(`<i data-role="hideMuted">`)
+              .attr('data-icon',lt.icon.BellOutline)
               .attr('title', trans.noarg('hideMutedTitle'))
               .on('click', () => {
                 $('main').toggleClass('lichessTools-hideMuted');
                 this.scrollIfNeeded();
-              }))
-          .appendTo(header);
+              }));
         }
       }
       const watchGamesTitle = trans.noarg('watchGames');
@@ -454,19 +461,39 @@
         const user = this.getUserId(m && m[1]);
         if (!user) return;
         this.rows[user] = row;
-        if (!actions.find('a.lichessTools-tv')[0]) {
-          $(`<a class="text lichessTools-tv" data-icon="${lt.icon.toEntity(lt.icon.AnalogTv)}"></a>`)
+        if (!isBlocks && !actions.find('a.lichessTools-tv')[0]) {
+          $(`<a class="text lichessTools-tv">`)
+            .attr('data-icon',lt.icon.AnalogTv)
             .attr('href', '/@/' + user + '/tv')
             .attr('title', watchGamesTitle)
             .prependTo(actions);
         }
         if (!isFavoritesOrBlocksOrFollowers && hasAlerts && !actions.find('a.lichessTools-mute')[0]) {
-          $(`<a class="text lichessTools-mute" data-icon="${lt.icon.toEntity(lt.icon.BellOutline)}"></a>`)
+          $(`<a class="text lichessTools-mute">`)
+            .attr('data-icon',lt.icon.BellOutline)
             .attr('title', mutePlayingAlertTitle)
             .on('click', ev => {
               ev.preventDefault();
               lt.pubsub.emit('lichessTools.mutePlayer', user);
               this.updateFriendsPage();
+            })
+            .appendTo(actions);
+        }
+        if (isBlocks && !actions.find('a.lichessTools-unblock')[0]) {
+          $(`<a class="text lichessTools-unblock">`)
+            .attr('data-icon',lt.icon.NotAllowed)
+            .attr('title', trans.noarg('unblockPlayerTitle'))
+            .attr('href','#')
+            .on('click', async (ev) => {
+              ev.preventDefault();
+              const link = $(ev.currentTarget);
+              if (link.prop('disabled')) {
+                return;
+              }
+              lt.api.relations.unblockPlayer(user);
+              $(ev.currentTarget)
+                .addClass('disabled')
+                .prop('disabled',true);
             })
             .appendTo(actions);
         }