+++ HISTORY.md
@@ -1,5 +1,9 @@
 History of features added to LiChess Tools in time (reversed order)
 
+02 Dec
+
+- unblock action in the blocked player page
+
 29 Nov
 
 - game analysis layout